+++ src/Resources/FilamentLatexResource/Pages/CreateFilamentLatex.php
@@ -4,6 +4,7 @@ namespace TheThunderTurner\FilamentLatex\Resources\FilamentLatexResource\Pages;
 
 use Filament\Resources\Pages\CreateRecord;
 use Illuminate\Contracts\Support\Htmlable;
+use Illuminate\Support\Facades\Storage;
 use TheThunderTurner\FilamentLatex\Resources\FilamentLatexResource;
 
 class CreateFilamentLatex extends CreateRecord
@@ -26,5 +27,11 @@ class CreateFilamentLatex extends CreateRecord
     protected function getCreatedNotificationTitle(): ?string
     {
         return 'Created Document';
+    }
+
+    protected function afterCreate(): void
+    {
+        $defaultContent = '\\documentclass{article}\n\\begin{document}\n% Your content here\n\\end{document}';
+        Storage::disk(config('filament-latex.storage'))->put($this->record->id . '/main.tex', $defaultContent);
     }
 }